day12/test_springs.py
import unittest

from approvaltests import verify

from day12.springs import SpringConditionRecord
from day12.springs_printer import print_working_spring_arrangements_for_file


class TestGalaxies(unittest.TestCase):

    def test_should_print_working_spring_arrangements_for_example(self):
        verify(print_working_spring_arrangements_for_file('example'))

    def test_should_count_arrangements_1(self):
        record = SpringConditionRecord("???.###", [1, 1, 3])
        self.assertEqual(1, record.arrangements().arrangements_count)

    def test_should_count_arrangements_2(self):
        record = SpringConditionRecord("?.#??.??#?", [2, 1, 1])
        self.assertEqual(1, record.arrangements().arrangements_count)

    def test_should_print_working_spring_arrangements_for_input(self):
        verify(print_working_spring_arrangements_for_file('input'))
